feat: jwt, 주문 까지 영역 작성, 결제 창

+++ src/main/java/com/bootcamp/paymentdemo/domain/order/dto/CreateOrderResponse.java
@@ -1,8 +1,8 @@
 package com.bootcamp.paymentdemo.domain.order.dto;
-// TODO 결제 테스트용, 삭제 필요
+
 public record CreateOrderResponse(
         String orderId
-        , Long totalAmount
+        , Integer totalAmount
         , String orderNumber
 ) {
 }
